feat(compendium): reorganize to separate canonical and custom content
- Move canonical unit database to /compendium/units (was /units)
- Move custom units browser to /units (was /compendium/units)
- Consolidate construction rules into single page with anchor nav
- Update home page navigation cards for new structure
- Update nav bar to show 'My Units' for custom units
- Fix compendium hub to properly reflect canonical content
- Fix construction rules grid to prevent label text cutoff
